v0.1.1 Fix rarity generation logic, add Arsenal
In some cases, an exception was thrown when creating a rarity because
the rarity was not generated due to an unfulfilled condition.
It was:
`value <= currentRange and value <= currentRange + chance`
the second condition is not relevant if the chance > 0.
It became:
`currentRange <= value and value <= currentRange + chance`

The appearance of the Arsenal is associated with the need to store
information about the created artifacts in the future.

Artifacts are generated in full with all the parameters as it was
intended. Previously it was not possible to generate 'subStats'.